requirements/requirement.txt
amqp==2.5.2
anyjson==0.3.3
appnope==0.1.0
APScheduler==3.6.3
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
billiard==3.6.1.0
bleach==3.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0
django-appconf==1.0.3
django-clear-cache==0.3
django-extensions==2.2.5
django-import-export==2.0
django-widget-tweaks==1.4.5
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
kombu==4.6.7
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
odfpy==1.4.0
openpyxl==3.0.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
redis==3.3.11
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12
sqlparse==0.3.0
stripe==2.41.0
tablib==0.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
djangorestframework==3.11.0
plotly==4.5.4
